+++ travel-board-app/src/assets/css/Detailstyles.js
